kaizn/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.2
Django==4.1.13
django-debug-toolbar==4.3.0
django-redis==5.4.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.5.0
hiredis==2.3.2
pycparser==2.21
pymongo==3.12.3
PyMySQL==1.1.0
pytz==2024.1
redis==5.0.1
sqlparse==0.2.4
